Overflow en la solución de I.
Corrí el programa sobre la entrada y me estaban dando números negativos.
Arreglé la solución para que se le quitara el overlfow. :P

+++ 2013/inline elephant toast/Solution/inline.cc
@@ -3,10 +3,10 @@
 
 using namespace std;
 
-long C[60][60];
-long blank;
+long long C[60][60];
+long long blank;
 
-long comb(int n, int k)
+long long comb(int n, int k)
 {
     if(k==0)
         return 1;
